[WD-1676]  Rebase onto upstream v0.15.2
This commit and the one after it reapply the commits from
https://github.com/german-first-iteration/eudi-lib-android-wallet-core/tree/develop-0.14.0-7b-partial-tweaks

on top of upstream version 0.15.2 which corresponds to
commit 0f0494eb4016c995d6b1ed4d19e2dfb12f431e16.

One minor merge conflict has been resolved: Upstream added a
dependency on robolectric in wallet-core/build.gradle.kts
near where we had also made an addition. I've kept the addition
of the robolectric dependency, although we probably don't strictly
need it.