[WD-1676] Fix core/build.gradle.kts to support inclusion of this project as a submodule
I read somewhere that "alias(...) apply false" only works for the root project, not for a nested project. This is okay for upstream because they consume the core library as a binary. For us, it's beneficial to build core from source, since that makes debugging easier and also speeds up the development cycle (don't need to publish named versions).